services/ai-service/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
gunicorn==23.0.0
psycopg2-binary==2.9.10
redis==5.2.1
celery==5.4.0
openai==1.58.1
detoxify==0.5.2
numpy==1.26.4
spacy==3.7.6
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
dj-database-url==2.3.0
whitenoise==6.8.2
channels==4.2.0
channels-redis==4.2.1
django-redis==5.4.0
